Move file_state_from_configuration to file_state.rs
